Time: 12 ms (91.59%), Space: 16.2 MB (46.04%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -9,7 +9,7 @@ public:
             for(int j=i+1;j<n;j++){
                 if(j>i+1 && nums[j]==nums[j-1]) continue;
                 int k=j+1, l=n-1;
-                int left= target-nums[i];
+                long left= target-nums[i];
                 left-= nums[j];
                 while(k<l){
                     if(nums[k]+nums[l]<left) k++;
